// src/api/models/body/common.body.ts
export interface BodyGetGLJoulist_1 {
  fromDate: string;
  toDate: string;
  last2Yr: number;
  closeYear: string;
  exclOther: string;
  summDate: string;
  fromFibu: string;
  toFibu: string;
  sorttype: number;
  fromDept: number;
  cashflow: string;
}

export interface BodyGetGLMainAccount {
  caseType: number;
  char1: string;
  char2: string;
  int1: number;
  int2: number;
}

export interface BodyGetHTParam0 {
  casetype: number;
  inpParam: 558;
}